17.10h

+++ app/users/model/Users.php
@@ -210,7 +210,6 @@ class Users extends Model{
             
             if(empty(self::$params_url[id])){
                 echo "Укажите ID";
-                die;
             }
             else{
                     $stmt = self::$db->prepare("DELETE FROM `users_person_data` WHERE id=:id");
@@ -221,10 +220,15 @@ class Users extends Model{
         }
     
          public function DelUsersCardsSql() {
-            
-            $stmt = self::$db->prepare("DELETE FROM `users_cards` WHERE id=:id");
-            $stmt -> bindValue(":id", self::$params_url["id"]);
-            $result_query = $stmt->execute();
-            echo "Пользователь " . self::$params_url["id"] . " успешно удален.";
+             if(empty(self::$params_url[id])){
+                echo "Укажите ID";
+             }
+             
+             else{
+                        $stmt = self::$db->prepare("DELETE FROM `users_cards` WHERE id=:id");
+                     $stmt -> bindValue(":id", self::$params_url["id"]);
+                        $result_query = $stmt->execute();
+                        echo "Пользователь " . self::$params_url["id"] . " успешно удален.";
+             }
         }
 }
